More getters, still refinements needed
